don't kill the pid if it does not exist

+++ config/deploy.rb
@@ -44,7 +44,10 @@ namespace :deploy do
 
   task :restart_unicorn do
     on roles(:app), in: :groups, limit: 3, wait: 10 do
-      execute "kill -QUIT `cat #{current_path}/tmp/pids/unicorn.pid`"
+      if test("[ -f #{current_path}/tmp/pids/unicorn.pid ]")
+        execute "kill -QUIT `cat #{current_path}/tmp/pids/unicorn.pid`"
+      end
+
       execute "cd #{current_path} ; /opt/rbenv/shims/bundle exec unicorn -c /u/apps/craigslistpoetry-api_production/shared/config/unicorn.rb -E production -D"
     end
   end
